third_party/blink/web_tests/scrollingcoordinator/non-fast-scrollable-region-nested-expected.txt

The test ensures that non-fast scrollable regions nested in iframe are correctly offset by the iframe location.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nonFastScrollableRects.length is 2
PASS rectToString(nonFastScrollableRects[0]) is "[51, 102, 200, 200]"
PASS rectToString(nonFastScrollableRects[1]) is "[51, 402, 211, 211]"
PASS successfullyParsed is true

TEST COMPLETE
